框架代码基本完成，基本的接收数据分包完成，但是没有仔细测试；接收数据缓冲重新封装了CharVector，因为vector<char>太慢了

+++ include/IDispatcher.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "CommonHeader.h"
 
 namespace robin
 {
@@ -6,6 +7,6 @@ namespace robin
 	class IDispatcher
 	{
 	public:
-		virtual void onMessage(void * client, char *buf, unsigned long len) {};
+		virtual void onMessage(void * client, char *buf, ssize_t len) {};
 	};
 }
